Added inertial reference frame test and fixed calculations

+++ app/logic/math/linear.py
@@ -12,10 +12,10 @@ def quaternion_multiply(quaternion1, quaternion0):
         x1 * w0 + y1 * z0 - z1 * y0 + w1 * x0,
         -x1 * z0 + y1 * w0 + z1 * x0 + w1 * y0,
         x1 * y0 - y1 * x0 + z1 * w0 + w1 * z0
-    ])
+    ], dtype=float)
 
 def rotate_vector_by_quaternion(v, q):
-    u = np.array(q[1:4])
+    u = np.array(q[1:4], dtype=float)
 
     s = q[0]
